+++ core/lib/appMain.d.ts
@@ -1,0 +1,16 @@
+export declare class appMain {
+    private readonly bw;
+    private readonly dsp;
+    private readonly screenRX;
+    private readonly screenRY;
+    private readonly hInfo;
+    private constructor();
+    private isMovingWin;
+    private posMovingWin;
+    private delayWinPos;
+    private window;
+    openDevTools(): void;
+    private static ins;
+    static initRenderer(path_htm: string, o: object): Electron.BrowserWindow;
+}
+//# sourceMappingURL=appMain.d.ts.map