+++ core/lib/index.d.ts
@@ -28,5 +28,12 @@ export * from "./sn/TxtLayer";
 export * from "./sn/TxtStage";
 export * from "./sn/Variable";
 export * from "./app";
+export * from "./appMain";
 export * from "./web";
+import { HPROC } from "./preload";
+declare global {
+    interface Window {
+        to_app: HPROC;
+    }
+}
 //# sourceMappingURL=index.d.ts.map