+++ core/src/sn/SysApp.ts
@@ -5,128 +5,217 @@
 	http://opensource.org/licenses/mit-license.php
 ** ***** END LICENSE BLOCK ***** */
 
-import {SysNode} from "./SysNode";
-import {SysBase} from "./SysBase";
+import {SysNode} from './SysNode';
+import {SysBase} from './SysBase';
 import {CmnLib, getDateStr, argChk_Boolean, argChk_Num, uint} from './CmnLib';
 import {ITag, IHTag, IVariable, IData4Vari, IMain} from './CmnInterface';
 import {Main} from './Main';
 import {Application} from 'pixi.js';
+import {createHash} from 'crypto';
 
+import {HINFO} from '../preload';
+/*	//== new ==
+const {to_app} = window;
+*/	//== new ==
+
+
+	//== old ==
 import {remote} from 'electron';
 const app = remote.app;
-//import * as Store from 'electron-store';
 
-//import {createWriteStream, removeSync, ensureDirSync, createReadStream, readFileSync, readFile, existsSync, copySync} from 'fs-extra';
-	let createWriteStream: Function;
-	let removeSync: Function;
-	let ensureDirSync: Function;
-	let createReadStream: Function;
-	let readFileSync: Function;
-	let readFile: Function;
-	let existsSync: Function;
-	let copySync: Function;
-import {createHash} from 'crypto';
-//import {pack, extract} from 'tar-fs';
-	let pack: Function;
-	let extract: Function;
+const dsp = remote.screen.getPrimaryDisplay();
+const screenRX = dsp.size.width;
+const screenRY = dsp.size.height;
+
+import {HPROC} from '../preload';
+import {existsSync, copySync, removeSync, ensureDirSync, ensureFileSync, createWriteStream, createReadStream, readFileSync, readFile, writeFileSync, appendFile} from 'fs-extra';
+const Store = require('electron-store');
+import {pack, extract} from 'tar-fs';
+
+const to_app: HPROC = {
+	getInfo		: ()=> {return {
+		getAppPath	: app.getAppPath(),
+		isPackaged	: app.isPackaged,
+		downloads	: app.getPath('downloads'),
+		userData	: app.getPath('userData'),
+		getVersion	: String(app.getVersion()),
+		env			: process.env,
+		screenResolutionX	: screenRX,
+		screenResolutionY	: screenRY,
+	}},
+
+	existsSync	: async (path: string)=> existsSync(path),
+	copySync	: (path_from: string, path_to: string)=> copySync(path_from, path_to),
+	removeSync	: async (path: string)=> removeSync(path),
+	ensureDirSync	: async (path: string)=> ensureDirSync(path),
+	ensureFileSync	: async (path: string)=> ensureFileSync(path),
+	createWriteStream	: async (path: string)=> createWriteStream(path),
+	createReadStream	: async (path: string)=> createReadStream(path),
+	readFileSync: async (path: string)=> readFileSync(path, {encoding: 'utf8'}),
+	readFile	: (path: string, callback: (err: NodeJS.ErrnoException, data: Buffer)=> void)=> readFile(path, callback),
+	writeFileSync	: async (path: string, data: Buffer, o?: object)=> writeFileSync(path, data, o),
+	appendFile		: async (path: string, data: string, callback: (err: Error)=> void)=> appendFile(path, data, callback),
+
+	window	: (centering: boolean, x: number, y: number, w: number, h: number)=> window(centering, x, y, w, h),
+	isSimpleFullScreen	: async ()=> bw.isSimpleFullScreen(),
+	setSimpleFullScreen	: async (b: boolean)=> bw.setSimpleFullScreen(b),
+	win_close		: ()=> bw.close(),
+	win_setTitle	: (title: string)=> bw.setTitle(title),
+	win_setContentSize	: async (w: number, h: number)=> bw.setContentSize(w, h),
+	win_setSize			: async (w: number, h: number)=> bw.setSize(w, h),
+
+	openDevTools	: ()=> bw.webContents.openDevTools(),
+	win_ev_devtools_opened	: (fnc: ()=> void)=> bw.webContents.on('devtools-opened', fnc),
+
+	Store	: async (o: object)=> {st = new Store(o); return;},
+	flush	: async (o: object)=> {st.store = o; return;},
+	Store_isEmpty	: async ()=> st.size === 0,
+	Store_get	: async ()=> st.store,
+
+	tarFs_pack		: async (path: string)=> pack(path),
+	tarFs_extract	: async (path: string)=> extract(path),
+};
+
+let	st: any;
+
+
+let bw: Electron.BrowserWindow = remote.getCurrentWindow();
+function window(centering: boolean, x: number, y: number, w: number, h: number) {
+	if (centering) {
+		const s = bw.getPosition();
+		x = (screenRX - s[0]) *0.5;
+		y = (screenRY - s[1]) *0.5;
+	}
+	else {
+		if (x < 0) x = 0; else if (x > screenRX) x = 0;
+		if (y < 0) y = 0; else if (y > screenRY) y = 0;
+	}
+	bw.setPosition(x, y);
+	bw.setContentSize(w, h);
+		// 2019/07/06 Windowsでこれがないとどんどん縦に短くなる
+	const hz = bw.getContentSize()[1];
+	bw.setContentSize(w, h *2 -hz);
+		// 2019/07/14 setContentSize()したのにメニュー高さぶん勝手に削られた値にされる不具合ぽい動作への対応
+}
+
+bw.on('move', ()=> {
+	if (isMovingWin) return;
+	isMovingWin = true;
+	posMovingWin = bw.getPosition();
+	setTimeout(()=> delayWinPos(), 500);
+});
+
+let	isMovingWin		= false;
+let	posMovingWin	= [0, 0];
+function	delayWinPos() {
+	if (bw.isSimpleFullScreen()) return;
+
+	const p = bw.getPosition();
+	if (posMovingWin[0] !== p[0] || posMovingWin[1] !== p[1]) {
+		posMovingWin = p;
+		setTimeout(()=> delayWinPos(), 500);
+		return;
+	}
+	window(false, p[0], p[1], 0, 0);
+	isMovingWin = false;
+}
+	//== old ==
+
 
 export class SysApp extends SysNode {
 	constructor(hPlg = {}, arg = {cur: 'prj/', crypto: false, dip: ''}) {
-		super(hPlg, {...arg, cur: app.getAppPath().replace(/\\/g, '/') + (app.isPackaged ?'/doc/' :'/')+ arg.cur});
+		super(hPlg, {...arg, cur: ''});
 
-		globalThis.addEventListener('DOMContentLoaded', ()=>this.run(), {once: true, passive: true});
+		globalThis.addEventListener('DOMContentLoaded', async ()=> {
+			this.hInfo = await to_app.getInfo();
+			this.arg = {...arg, cur: this.hInfo.getAppPath.replace(/\\/g, '/') + (this.hInfo.isPackaged ?'/doc/' :'/')+ arg.cur};
 
-		(async ()=> {
-			const {ipcRenderer} = await import('electron');
-			ipcRenderer.on('log', (e: any, arg: any)=> console.log(`[main log] e:%o arg:%o`, e, arg));
+	//== old ==
+			this.readFileSync = to_app.readFileSync;
+			this.writeFileSync = to_app.writeFileSync;
+			this.appendFile = to_app.appendFile;
+			this.ensureFileSync = to_app.ensureFileSync;
+	//== old ==
 
-			const {createWriteStream: fe0, removeSync: fe1, ensureDirSync: fe2, createReadStream: fe3, readFileSync: fe4, readFile: fe5, existsSync: fe6, copySync: fe7} = await import('fs-extra');
-			createWriteStream = fe0;
-			removeSync = fe1;
-			ensureDirSync = fe2;
-			createReadStream = fe3;
-			readFileSync = fe4;
-			readFile = fe5;
-			existsSync = fe6;
-			copySync = fe7;
+			this.$path_downloads	= this.hInfo.downloads.replace(/\\/g, '/') +'/';
 
-			const {pack: pc, extract: ext} = await import('tar-fs');
-			pack = pc;
-			extract = ext;
-		})();
+//			ipcRenderer.on('log', (e: any, arg: any)=> console.log(`[main log] e:%o arg:%o`, e, arg));
+
+			this.run();
+		}, {once: true, passive: true});
 	}
-	protected 			$path_userdata	: string;
-	protected readonly	$path_downloads	= app.getPath('downloads').replace(/\\/g, '/') +'/';
+	private	hInfo:  HINFO = {
+		getAppPath	: '',
+		isPackaged	: false,
+		downloads	: '',
+		userData	: '',
+		getVersion	: '',
+		env			: {},
+		screenResolutionX	: 0,
+		screenResolutionY	: 0,
+	};
+
+/*	//== new ==
+	protected	readFileSync = to_app.readFileSync;
+	protected	writeFileSync = to_app.writeFileSync;
+	appendFile = to_app.appendFile;
+	ensureFileSync = to_app.ensureFileSync;
+*/	//== new ==
+
+	protected 	$path_userdata	= '';
+	protected	$path_downloads	= '';
 
 	protected readonly	normalize = (src: string, form: string)=> src.normalize(form);
 
 	initVal(data: IData4Vari, hTmp: any, comp: (data: IData4Vari)=> void) {
 		this.$path_userdata	= this.isDbg()
-			? app.getAppPath().slice(0, -3) +'.vscode/'	// /doc → /
-			: app.getPath('userData').replace(/\\/g, '/') +'/';
-		const Store = require('electron-store');
-		const st = new Store({
-			cwd: this.$path_userdata +'storage',
-			name: this.arg.crypto ?'data_' :'data',
-			encryptionKey: this.arg.crypto ?this.stk() :undefined,
-		});
-		this.flush = ()=> st.store = this.data;
+			? this.hInfo.getAppPath.slice(0, -3) +'.vscode/'	// /doc → /
+			: this.hInfo.userData.replace(/\\/g, '/') +'/';
+		(async()=> {
+			await to_app.Store({
+				cwd: this.$path_userdata +'storage',
+				name: this.arg.crypto ?'data_' :'data',
+				encryptionKey: this.arg.crypto ?this.stk() :undefined,
+			});
+			this.flush = ()=> to_app.flush(this.data);
 
-		if (hTmp['const.sn.isFirstBoot'] = (st.size === 0)) {
-			// データがない（初回起動）場合の処理
-			this.data.sys = data.sys;
-			this.data.mark = data.mark;
-			this.data.kidoku = data.kidoku;
-			this.flush();	// 初期化なのでここのみ必要
-		}
-		else {
-			// データがある場合の処理
-			this.data.sys = st.store.sys as any;
-			this.data.mark = st.store.mark as any;
-			this.data.kidoku = st.store.kidoku as any;
-		}
-		comp(this.data);
+			if (hTmp['const.sn.isFirstBoot'] = await to_app.Store_isEmpty()) {
+				// データがない（初回起動）場合の処理
+				this.data.sys = data.sys;
+				this.data.mark = data.mark;
+				this.data.kidoku = data.kidoku;
+				await this.flush();	// 初期化なのでここのみ必要
+			}
+			else {
+				// データがある場合の処理
+				const store = await to_app.Store_get();
+				this.data.sys = store.sys;
+				this.data.mark = store.mark;
+				this.data.kidoku = store.kidoku;
+			}
+			comp(this.data);
+		})();
 
 		// システム情報
 		hTmp['const.sn.isDebugger'] = false;
 			// システムがデバッグ用の特別なバージョンか
 			// AIRNovel の const.flash.system.Capabilities.isDebugger
-		hTmp['const.sn.screenResolutionX'] = this.dsp.size.width;
+		hTmp['const.sn.screenResolutionX'] = this.hInfo.screenResolutionX;
 			// 画面の最大水平解像度
-		hTmp['const.sn.screenResolutionY'] = this.dsp.size.height;
+		hTmp['const.sn.screenResolutionY'] = this.hInfo.screenResolutionY;
 			// 画面の最大垂直解像度
 			// AIRNovel の const.flash.system.Capabilities.screenResolutionX、Y
 			// 上のメニューバーは含んでいない（たぶん an も）。含むのは workAreaSize
 
-		this.val.defTmp('const.sn.displayState', ()=> this.win.isSimpleFullScreen());
+		this.val.defTmp('const.sn.displayState', async ()=> await to_app.isSimpleFullScreen());
 
+/*
 		globalThis.addEventListener('resize', ()=> {
 			// NOTE: 2019/07/14 Windowsでこのように遅らせないと正しい縦幅にならない
 			this.window((hTmp['const.sn.isFirstBoot']) ?{centering: true} :{});
 		}, {once: true, passive: true});
-
-		this.win.on('move', ()=> {
-			if (this.isMovingWin) return;
-			this.isMovingWin = true;
-			this.posMovingWin = this.win.getPosition();
-			setTimeout(()=> this.delayWinPos(), 500);
-		});
+*/
 	}
-	private	isMovingWin	= false;
-	private posMovingWin= [0, 0];
-	private delayWinPos() {
-		if (this.win.isSimpleFullScreen()) return;
-
-		const p = this.win.getPosition();
-		if (this.posMovingWin[0] !== p[0] || this.posMovingWin[1] !== p[1]) {
-			this.posMovingWin = p;
-			setTimeout(()=> this.delayWinPos(), 500);
-			return;
-		}
-		this.window({x: p[0], y: p[1]});
-		this.isMovingWin = false;
-	}
-	private readonly	dsp	= remote.screen.getPrimaryDisplay();
 
 
 	private main: Main;
@@ -141,53 +230,53 @@ export class SysApp extends SysNode {
 	}
 
 
-	private readonly	win	= remote.getCurrentWindow();
-	private readonly	wc	= this.win.webContents;
 	init(hTag: IHTag, appPixi: Application, val: IVariable, main: IMain) {
 		super.init(hTag, appPixi, val, main);
 
-		if (this.cfg.oCfg.debug.devtool) this.wc.openDevTools();
-		else this.wc.on('devtools-opened', ()=> {
+		if (this.cfg.oCfg.debug.devtool) to_app.openDevTools();
+		else to_app.win_ev_devtools_opened(()=> {
 			console.error(`DevToolは禁止されています。許可する場合は【プロジェクト設定】の【devtool】をONに。`);
 			main.destroy();
 		});
-		this.win.setContentSize(CmnLib.stageW, CmnLib.stageH);
+		to_app.win_setContentSize(CmnLib.stageW, CmnLib.stageH);
 	}
 
 
-	copyBMFolder = (from: number, to: number)=> {
+	copyBMFolder = async (from: number, to: number)=> {
 		const path_from = `${this.$path_userdata}storage/${from}/`;
 		const path_to = `${this.$path_userdata}storage/${to}/`;
-		if (! existsSync(path_from)) return;	// 使ってない場合もある
+		if (! await to_app.existsSync(path_from)) return;	// 使ってない場合もある
 
-		copySync(path_from, path_to);
+		await to_app.copySync(path_from, path_to);
 	};
-	eraseBMFolder = (place: number)=> {
-		removeSync(`${this.$path_userdata}storage/${place}/`);
+	eraseBMFolder = async (place: number)=> {
+		await to_app.removeSync(`${this.$path_userdata}storage/${place}/`);
 	};
 
-	protected readonly	isPackaged = ()=> app.isPackaged;
+	protected readonly	isPackaged = ()=> this.hInfo.isPackaged;
 	isDbg = ()=> {	// 配布版では無効
-		const ret = Boolean(process.env['SKYNOVEL_DBG']) && ! this.isPackaged();
+		const ret = Boolean(this.hInfo.env['SKYNOVEL_DBG']) && ! this.isPackaged();
 		this.isDbg = ret ?()=> true :()=> false;
-		if (ret) this.extPort = uint(process.env['SKYNOVEL_PORT'] ?? '3776');
+		if (ret) this.extPort = uint(this.hInfo.env['SKYNOVEL_PORT'] ?? '3776');
 		return ret;
 	};
 
 	// アプリの終了
-	protected readonly	close = ()=> {this.win.close(); return false;}
+	protected readonly	close = ()=> {to_app.win_close(); return false;}
 
 	// プレイデータをエクスポート
 	protected readonly	_export: ITag = ()=> {
-		const r = pack(this.$path_userdata +'storage/')
-		r.on('end', ()=> {
-			if (CmnLib.debugLog) console.log('プレイデータをエクスポートしました');
-			this.fire('sn:exported', new Event('click'));
-		});
-		r.pipe(createWriteStream(
-			this.$path_downloads + (this.crypto ?'' :'no_crypto_')
-			+ this.cfg.getNs() + getDateStr('-', '_', '') +'.spd'
-		));
+		(async()=> {
+			const r = await to_app.tarFs_pack(this.$path_userdata +'storage/')
+			r.on('end', ()=> {
+				if (CmnLib.debugLog) console.log('プレイデータをエクスポートしました');
+				this.fire('sn:exported', new Event('click'));
+			});
+			r.pipe(await to_app.createWriteStream(
+				this.$path_downloads + (this.crypto ?'' :'no_crypto_')
+				+ this.cfg.getNs() + getDateStr('-', '_', '') +'.spd'
+			));
+		})();
 
 		return false;
 	}
@@ -202,16 +291,16 @@ export class SysApp extends SysNode {
 			inp.onchange = ()=> {if (inp.files) rs(inp.files[0].path); else rj();};
 			inp.click();
 		})
-		.then((inp_path: any)=> {
+		.then(async (inp_path: any)=> {
 			this.flush = ()=> {};
 			const out_path = this.$path_userdata +'storage/';
-			removeSync(out_path);
-			ensureDirSync(out_path);	// ディレクトリ、なければ作る
+			await to_app.removeSync(out_path);
+			await to_app.ensureDirSync(out_path);	// ディレクトリ、なければ作る
 
-			createReadStream(inp_path)
+			(await to_app.createReadStream(inp_path))
 			.on('end', async ()=> {
 				const fn = this.$path_userdata +'storage/data.json'+ (this.crypto ?'_': '');
-				const s = String(readFileSync(fn));
+				const s = await to_app.readFileSync(fn);
 				const o = JSON.parse(this.crypto ?await this.pre('json', s) :s);
 				if (! o.sys || ! o.mark || ! o.kidoku) throw new Error('異常なプレイデータです');
 				if (o.sys[SysBase.VALNM_CFG_NS] !== this.cfg.oCfg.save_ns) {
@@ -223,13 +312,13 @@ export class SysApp extends SysNode {
 				this.data.mark = o.mark;
 				this.data.kidoku = o.kidoku;
 				this.flush = flush;
-				this.flush();
+				await this.flush();
 				this.val.updateData(o);
 
 				if (CmnLib.debugLog) console.log('プレイデータをインポートしました');
 				this.fire('sn:imported', new Event('click'));
 			})
-			.pipe(extract(out_path));
+			.pipe(await to_app.tarFs_extract(out_path));
 		});
 
 		return false;
@@ -249,7 +338,7 @@ export class SysApp extends SysNode {
 		return false;
 	}
 	// タイトル指定
-	protected titleSub(txt: string) {this.win.setTitle(txt);}
+	protected titleSub(title: string) {to_app.win_setTitle(title);}
 	// 全画面状態切替
 	protected readonly	tgl_full_scr: ITag = hArg=> {
 		if (! hArg.key) {this.tgl_full_scr_sub(); return false;}
@@ -267,10 +356,10 @@ export class SysApp extends SysNode {
 		}, {passive: true});
 		return false;
 	}
-	protected readonly	tgl_full_scr_sub = ()=> {
-		if (this.win.isSimpleFullScreen()) {
-			this.win.setSimpleFullScreen(false);	// これはこの位置
-			this.win.setSize(CmnLib.stageW, CmnLib.stageH);
+	protected readonly	tgl_full_scr_sub = async ()=> {
+		if (await to_app.isSimpleFullScreen()) {
+			await to_app.setSimpleFullScreen(false);	// これはこの位置
+			await to_app.win_setSize(CmnLib.stageW, CmnLib.stageH);
 			this.appPixi.view.style.width  = CmnLib.stageW +'px';
 			this.appPixi.view.style.height = CmnLib.stageH +'px';
 			this.appPixi.view.style.marginLeft = '0px';
@@ -280,12 +369,12 @@ export class SysApp extends SysNode {
 			this.reso4frame = 1;
 		}
 		else {
-			const w = this.dsp.size.width;
-			const h = this.dsp.size.height;
+			const w = this.hInfo.screenResolutionX;
+			const h = this.hInfo.screenResolutionY;
 			const ratioWidth  = w / CmnLib.stageW;
 			const ratioHeight = h / CmnLib.stageH;
 			const ratio = (ratioWidth < ratioHeight) ?ratioWidth :ratioHeight;
-			this.win.setSize(CmnLib.stageW * ratio, CmnLib.stageH * ratio);
+			await to_app.win_setSize(CmnLib.stageW * ratio, CmnLib.stageH * ratio);
 			this.appPixi.view.style.width  = (CmnLib.stageW * ratio) +'px';
 			this.appPixi.view.style.height = (CmnLib.stageH * ratio) +'px';
 			if (ratioWidth < ratioHeight) {	// 左に寄る対策
@@ -294,9 +383,9 @@ export class SysApp extends SysNode {
 			else {
 				this.appPixi.view.style.marginLeft= (w -CmnLib.stageW *ratio) /2 +'px';
 			}
-			this.win.setSimpleFullScreen(true);	// これはこの位置
+			await to_app.setSimpleFullScreen(true);	// これはこの位置
 
-			this.win.setContentSize(screen.width, screen.height);
+			await to_app.win_setContentSize(screen.width, screen.height);
 				// これがないとWinアプリ版で下部が短くなり背後が見える
 			const cr = this.appPixi.view.getBoundingClientRect();
 			this.reso4frame = cr.width / CmnLib.stageW;
@@ -318,7 +407,7 @@ export class SysApp extends SysNode {
 			if (! mv) throw `[update_check] ファイル内にversionが見つかりません`;
 			const netver = mv[1];
 
-			const myver = String(app.getVersion());
+			const myver = this.hInfo.getVersion;
 			if (netver === myver) {
 				if (CmnLib.debugLog) console.log(`[update_check] バージョン更新なし ver:${myver}`);
 				return;
@@ -327,15 +416,15 @@ export class SysApp extends SysNode {
 
 			const o = {
 				title: 'アプリ更新',
-				icon: app.getAppPath() +'/app/icon.png',
+				icon: this.hInfo.getAppPath +'/app/icon.png',
 				buttons: ['OK', 'Cancel'],
 				defaultId: 0,
 				cancelId: 1,
 				message: `アプリ【${this.cfg.oCfg.book.title}】に更新があります。\nダウンロードしますか？`,
 				detail: `現在ver ${myver}\n新規ver ${netver}`,
 			};
-			const di = await remote.dialog.showMessageBox(o as any);// iconのせい
-			if (di.response > 0) return;
+//			const di = await remote.dialog.showMessageBox(o as any);// iconのせい
+//			if (di.response > 0) return;
 
 			if (CmnLib.debugLog) console.log(`[update_check] アプリダウンロード開始`);
 			const mp = /path: (.+)/.exec(txt);
@@ -348,7 +437,7 @@ export class SysApp extends SysNode {
 
 			const res_dl = await this.fetch(url + fn);
 			if (! res_dl.ok) return;
-			const pathDL = app.getPath('downloads') +'/'+ fn;
+			const pathDL = this.hInfo.downloads +'/'+ fn;
 			const rd_dl = async (res: Response)=> {
 				const reader = res!.body!.getReader();
 				const {Readable} = await import('stream');
@@ -364,7 +453,7 @@ export class SysApp extends SysNode {
 			pipe_dl.on('end', ()=> {
 				if (CmnLib.debugLog) console.log(`[update_check] アプリダウンロード完了`);
 
-				readFile(pathDL, (err: any, data: any)=> {
+				to_app.readFile(pathDL, async (err: any, data: any)=> {
 					if (err) throw err;
 
 					const h = createHash('SHA512');
@@ -373,44 +462,28 @@ export class SysApp extends SysNode {
 
 					const isOk = sha === hash;
 					if (CmnLib.debugLog) console.log(`[update_check] SHA512 Checksum:${isOk}`, sha, hash);
-					if (! isOk) removeSync(pathDL);
+					if (! isOk) await to_app.removeSync(pathDL);
 
 					o.buttons.pop();
 					o.message = `アプリ【${this.cfg.oCfg.book.title}】の更新パッケージを\nダウンロードしました`+ (isOk ?'' :'が、破損しています。\n開発元に連絡してください');
-					remote.dialog.showMessageBox(o as any);	// iconのせい
+//					remote.dialog.showMessageBox(o as any);	// iconのせい
 				});
 			});
-			pipe_dl.pipe(createWriteStream(pathDL));
+			pipe_dl.pipe(await to_app.createWriteStream(pathDL));
 		})();
 
 		return false;
 	}
 	// アプリウインドウ設定
 	protected readonly	window: ITag = hArg=> {
-		const screenRX = this.dsp.size.width;
-		const screenRY = this.dsp.size.height;
-		if (argChk_Boolean(hArg, 'centering', false)) {
-			const s = this.win.getPosition();
-			hArg.x = (screenRX - s[0]) *0.5;
-			hArg.y = (screenRY - s[1]) *0.5;
-		}
-		else {
-			hArg.x = argChk_Num(hArg, 'x', Number(this.val.getVal('sys:const.sn.nativeWindow.x', 0)));
-			hArg.y = argChk_Num(hArg, 'y', Number(this.val.getVal('sys:const.sn.nativeWindow.y', 0)));
-			if (hArg.x < 0) hArg.x = 0;
-			else if (hArg.x > screenRX) hArg.x = 0;
-			if (hArg.y < 0) hArg.y = 0;
-			else if (hArg.y > screenRY) hArg.y = 0;
-		}
-		this.win.setPosition(hArg.x, hArg.y);
-		this.win.setContentSize(CmnLib.stageW, CmnLib.stageH);
-			// 2019/07/06 Windowsでこれがないとどんどん縦に短くなる
-		const hz = this.win.getContentSize()[1];
-		this.win.setContentSize(CmnLib.stageW, CmnLib.stageH *2 -hz);
-			// 2019/07/14 setContentSize()したのにメニュー高さぶん勝手に削られた値にされる不具合ぽい動作への対応
-		this.val.setVal_Nochk('sys', 'const.sn.nativeWindow.x', hArg.x);
-		this.val.setVal_Nochk('sys', 'const.sn.nativeWindow.y', hArg.y);
-		this.flush();
+		(async()=> {
+			const x = argChk_Num(hArg, 'x', Number(this.val.getVal('sys:const.sn.nativeWindow.x', 0)));
+			const y = argChk_Num(hArg, 'y', Number(this.val.getVal('sys:const.sn.nativeWindow.y', 0)));
+			await to_app.window(argChk_Boolean(hArg, 'centering', false), x, y, CmnLib.stageW, CmnLib.stageH);
+			this.val.setVal_Nochk('sys', 'const.sn.nativeWindow.x', x);
+			this.val.setVal_Nochk('sys', 'const.sn.nativeWindow.y', y);
+			await this.flush();
+		})();
 
 		return false;
 	}